# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pompote-view API",
    "description": "Read-only HTTP surface for monitoring trading strategies. Every `/api/v1/*` route requires the `X-API-Key` header.",
    "contact": {
      "name": "Pierre Attard"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/monitoring/backtests/{run_id}/candles": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/backtests/{run_id}/candles`.",
        "operationId": "get_backtest_candles",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "description": "Backtest run identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Aggregation width (must be one of [`Timeframe::ALLOWED`]).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated candles_5s over the run's market window, with a `source` hint.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestCandlesDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeframe or window exceeding the bucket cap.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "404": {
            "description": "No run with that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
...
    "/api/v1/monitoring/candles": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/candles`.",
        "operationId": "get_candles",
        "parameters": [
          {
            "name": "exchange",
            "in": "query",
            "description": "Exchange identifier as stored in `candles_5s.exchange`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "symbol",
            "in": "query",
            "description": "Symbol identifier as stored in `candles_5s.symbol`.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "timeframe",
            "in": "query",
            "description": "Aggregation width (must be one of [`Timeframe::ALLOWED`]).",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `open_time` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `open_time` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aggregated OHLCV buckets ordered by ascending `ts`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandleDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid timeframe, range or too many points.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error (e.g. schema drift).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandleErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/strategies": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/strategies`.",
        "operationId": "list_strategies",
        "responses": {
          "200": {
            "description": "Every strategy, ordered by name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StrategyDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
    "/api/v1/monitoring/strategies/{id}/decisions": {
      "get": {
        "tags": [
          "monitoring"
        ],
        "summary": "Handler for `GET /api/v1/monitoring/strategies/{id}/decisions`.",
        "operationId": "get_strategy_decisions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Strategy identifier",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "from",
            "in": "query",
            "description": "Inclusive lower bound on `executed_at` (RFC3339).",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "Exclusive upper bound on `executed_at` (RFC3339). Defaults to `Clock::now()`.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Row cap (defaults to [`MAX_ORDER_ROWS`], rejected above it).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Live decisions (with market context) for the strategy on the window, ordered by ascending `created_at`.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LiveDecisionDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid range or limit out of bounds.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `X-API-Key` header."
          },
          "500": {
            "description": "Unexpected internal error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "Datastore temporarily unreachable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "x_api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "BacktestCandlesDto": {
        "type": "object",
        "description": "Response for the run candles endpoint: the OHLC series plus a source hint.",
        "required": [
          "source",
          "candles"
        ],
        "properties": {
          "candles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandleDto"
            },
            "description": "Aggregated OHLC buckets (empty when `source = none`)."
          },
          "source": {
            "type": "string",
            "description": "`candles_5s` when the run window is covered, `none` for the degraded\ntimeline-only view (no OHLC background available).",
            "example": "candles_5s"
          }
        }
      },
      "BacktestErrorBody": {
        "type": "object",
        "description": "JSON error body for the backtest endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "allowed": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Whitelisted values for `invalid_status`."
          },
          "error": {
            "type": "string",
            "description": "Machine-readable discriminator (e.g. `invalid_range`)."
          },
          "max": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Applicable row cap for `too_many_rows`.",
            "minimum": 0
          },
          "message": {
            "type": [
              "string",
              "null"
            ],
            "description": "Human-readable explanation."
          },
          "requested": {
            "type": [
              "integer",
              "null"
            ],
            "description": "Echoes the requested row count for `too_many_rows`.",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "x_api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}